+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 ## [Unreleased]
 
+- Nothing at the moment
+
+## [2.1.0] - 2020-02-15
+
 ### Added
 
 - Insert always-zero tax line item in total breakdown
@@ -56,7 +60,8 @@ Initial release
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/2.0.0...master
+[Unreleased]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/2.1.0...master
+[2.1.0]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/2.0.0...2.1.0
 [2.0.0]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/1.0.0...2.0.0
 [1.0.0]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/0.1.0...1.0.0
 [0.1.0]: https://nosuchdomain.mooo.com/git/doc/aliexpress-invoice-generator/compare/0.0.1...0.1.0